W172a: mdlFile two window clears via WIDTH recovery (over-window 54 -> 52). 003579B0 508/496 nd374 -> 488/496 (parameter/return widths corrected to u32) and 0034BFC0 436/416 -> 416/416 (allocator result and return widths to u32). 003461A0 468/432 nd294 -> 448/432 nd264 by the same width correction, and that corrected signature TRANSITIVELY improved its caller 003463C0 156/160 nd76 -> 152/160 nd21 -- a good reminder that fixing a callee's real ABI pays at every call site. 0033A220 448/448 nd73 -> 57 by removing a reusable packed-byte pointer in favour of base-first constant-offset accesses. Zero lost, zero additions, 667 markers unchanged, push/pop 3/3 and opt_loop_invariants 16/16.

+++ src/Battle/btlFormation.c
@@ -13644,7 +13644,7 @@ u32 func_002cec30(void)
   extern void func_0035f060(u32);
   u32 unit;
   u32 selector;
-  s32 count;
+  u16 count;
   u16 index;
   int selected;
   int *candidate;
